修复long，integer的反序列化转换错误

+++ autofull-spring-boot-starter-example/src/main/java/com/suke/autofull/example/entity/TbMenu.java
@@ -19,7 +19,7 @@ import java.io.Serializable;
 public class TbMenu implements Serializable {
 
     @TableId(type = IdType.AUTO)
-    public int menuId;
+    public Long menuId;
 
     public String menuName;
 }
